refactor: move `gen` and `install` cmds to `sources` subdomain
