+++ cmd/gen.go
@@ -1,18 +1,8 @@
 package cmd
 
 import (
-	"github.com/metafates/mangal/constant"
-	"github.com/metafates/mangal/filesystem"
-	"github.com/metafates/mangal/util"
-	"github.com/metafates/mangal/where"
 	"github.com/samber/lo"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-	"os"
-	"os/user"
-	"path/filepath"
-	"strings"
-	"text/template"
 )
 
 func init() {
@@ -26,56 +16,9 @@ func init() {
 }
 
 var genCmd = &cobra.Command{
-	Use:   "gen",
-	Short: "Generate a new lua source",
-	Long:  `Generate a new lua source.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		cmd.SetOut(os.Stdout)
-
-		author := viper.GetString(constant.GenAuthor)
-		if author == "" {
-			usr, err := user.Current()
-			if err == nil {
-				author = usr.Username
-			} else {
-				author = "Anonymous"
-			}
-		}
-
-		s := struct {
-			Name            string
-			URL             string
-			SearchMangaFn   string
-			MangaChaptersFn string
-			ChapterPagesFn  string
-			Author          string
-		}{
-			Name:            lo.Must(cmd.Flags().GetString("name")),
-			URL:             lo.Must(cmd.Flags().GetString("url")),
-			SearchMangaFn:   constant.SearchMangaFn,
-			MangaChaptersFn: constant.MangaChaptersFn,
-			ChapterPagesFn:  constant.ChapterPagesFn,
-			Author:          author,
-		}
-
-		funcMap := template.FuncMap{
-			"repeat": strings.Repeat,
-			"plus":   func(a, b int) int { return a + b },
-			"max":    util.Max[int],
-		}
-
-		tmpl, err := template.New("source").Funcs(funcMap).Parse(constant.SourceTemplate)
-		handleErr(err)
-
-		target := filepath.Join(where.Sources(), util.SanitizeFilename(s.Name)+".lua")
-		f, err := filesystem.Api().Create(target)
-		handleErr(err)
-
-		defer util.Ignore(f.Close)
-
-		err = tmpl.Execute(f, s)
-		handleErr(err)
-
-		cmd.Println(target)
-	},
+	Use:        "gen",
+	Short:      "Generate a new lua source",
+	Long:       `Generate a new lua source.`,
+	Deprecated: "use `mangal sources gen` instead.",
+	Run:        sourcesGenCmd.Run,
 }
