+++ cmd/sources.go
@@ -2,8 +2,15 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/metafates/mangal/constant"
+	"github.com/metafates/mangal/tui"
+	"github.com/metafates/mangal/util"
+	"github.com/spf13/viper"
 	"os"
+	"os/user"
 	"path/filepath"
+	"strings"
+	"text/template"
 
 	"github.com/metafates/mangal/filesystem"
 	"github.com/metafates/mangal/icon"
@@ -73,5 +80,84 @@ var sourcesRemoveCmd = &cobra.Command{
 			handleErr(filesystem.Api().Remove(path))
 			fmt.Printf("%s successfully removed %s\n", icon.Get(icon.Success), style.Yellow(name))
 		}
+	},
+}
+
+func init() {
+	sourcesCmd.AddCommand(sourcesInstallCmd)
+}
+
+var sourcesInstallCmd = &cobra.Command{
+	Use:   "install",
+	Short: "Browse and install custom scrapers",
+	Long: `Browse and install custom scrapers from official GitHub repo.
+https://github.com/metafates/mangal-scrapers`,
+	Run: func(cmd *cobra.Command, args []string) {
+		handleErr(tui.Run(&tui.Options{Install: true}))
+	},
+}
+
+func init() {
+	sourcesCmd.AddCommand(sourcesGenCmd)
+
+	sourcesGenCmd.Flags().StringP("name", "n", "", "name of the source")
+	sourcesGenCmd.Flags().StringP("url", "u", "", "url of the website")
+
+	lo.Must0(sourcesGenCmd.MarkFlagRequired("name"))
+	lo.Must0(sourcesGenCmd.MarkFlagRequired("url"))
+}
+
+var sourcesGenCmd = &cobra.Command{
+	Use:   "gen",
+	Short: "Generate a new lua source",
+	Long:  `Generate a new lua source.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		cmd.SetOut(os.Stdout)
+
+		author := viper.GetString(constant.GenAuthor)
+		if author == "" {
+			usr, err := user.Current()
+			if err == nil {
+				author = usr.Username
+			} else {
+				author = "Anonymous"
+			}
+		}
+
+		s := struct {
+			Name            string
+			URL             string
+			SearchMangaFn   string
+			MangaChaptersFn string
+			ChapterPagesFn  string
+			Author          string
+		}{
+			Name:            lo.Must(cmd.Flags().GetString("name")),
+			URL:             lo.Must(cmd.Flags().GetString("url")),
+			SearchMangaFn:   constant.SearchMangaFn,
+			MangaChaptersFn: constant.MangaChaptersFn,
+			ChapterPagesFn:  constant.ChapterPagesFn,
+			Author:          author,
+		}
+
+		funcMap := template.FuncMap{
+			"repeat": strings.Repeat,
+			"plus":   func(a, b int) int { return a + b },
+			"max":    util.Max[int],
+		}
+
+		tmpl, err := template.New("source").Funcs(funcMap).Parse(constant.SourceTemplate)
+		handleErr(err)
+
+		target := filepath.Join(where.Sources(), util.SanitizeFilename(s.Name)+".lua")
+		f, err := filesystem.Api().Create(target)
+		handleErr(err)
+
+		defer util.Ignore(f.Close)
+
+		err = tmpl.Execute(f, s)
+		handleErr(err)
+
+		cmd.Println(target)
 	},
 }
